% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{check_qc_data}
\alias{check_qc_data}
\title{check_qc_data}
\usage{
check_qc_data(
  path = ".",
  column = c("c18", "hilic"),
  polarity = c("positive", "negative")
)
}
\arguments{
\item{path}{working path. It should contain two folders: "poolQC", "ms2". "poolQC" contains}

\item{column}{'c18', 'hilic'}

\item{polarity}{"positive", 'negative'}
}
\description{
check_qc_data
}
\author{
Zhiwei Zhou
}
