% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{get_summary_biomet}
\alias{get_summary_biomet}
\title{get_summary_biomet}
\usage{
get_summary_biomet(object, column, polarity, mz_ppm = 10, rt_second = 25)
}
\arguments{
\item{object}{tidymass object}

\item{column}{'hilic',  'c18'}

\item{polarity}{"positive", 'negative'}

\item{mz_ppm}{10 ppm in defalut}

\item{rt_second}{25 s in defalut}
}
\description{
get_summary_biomet
}
\examples{
load('~/Project/00_IBD_project/Data/20230808_data_quality_check/B0002/hilic_pos/object.RData')
get_summary_biomet(object = object, polarity = 'positive', column = 'hilic')
}
\author{
Zhiwei Zhou
}
