% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{plot_ms2_biomet}
\alias{plot_ms2_biomet}
\title{plot_ms2_biomet}
\usage{
plot_ms2_biomet(variable_id, metabolite_name, annot_table, ms2_result)
}
\arguments{
\item{annot_table}{intermidate data from DoddLabMetID}

\item{ms2_result}{intermidate data from DoddLabMetID}
}
\description{
plot_ms2_biomet
}
\examples{
variable_id <-  'M166T279'
metabolite_name <- 'L-Phenylalanine'
plot_ms2_biomet(variable_id = 'M166T279',
                metabolite_name = 'L-Phenylalanine',
                annot_table = annot_table,
                ms2_result = ms2_result)
}
\author{
Zhiwei Zhou
}
