% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.R
\name{process_qc_data}
\alias{process_qc_data}
\title{process_qc_data}
\usage{
process_qc_data(
  path = ".",
  column = c("c18", "hilic"),
  polarity = c("positive", "negative")
)
}
\arguments{
\item{path}{'.'}

\item{column}{'c18', 'hilic'. Default: 'c18'}

\item{polarity}{'positive', 'negative'. Default: "positive"}
}
\description{
process_qc_data
}
\author{
Zhiwei Zhou
}
